requirements.txt
flask
Pillow
opencv-python
numpy